Evaluate checkout ref when determining if a checkout is local
When a workflow tries to run actions/checkout, the runner needs to know if the
workflow wants to check out the current repository or a remote repository. It
does this by comparing the provided repository name and ref with those of the
current branch. This commit expands expressions in those fields before the
comparison, so workflows can call actions/checkout with parameters like
"ref: {{ SOME_VARIABLE }}".